+++ rs/src/cfg_actions.rs
@@ -28,29 +28,5 @@ pub struct CfgDefineFunction2 {
 #[derive(Serialize, Deserialize, Debug, PartialEq, Eq)]
 #[serde(rename_all = "snake_case")]
 pub struct CfgIf {
-  pub target: super::CfgLabel,
-}
-
-#[derive(Serialize, Deserialize, Debug, PartialEq, Eq)]
-#[serde(rename_all = "snake_case")]
-pub struct CfgJump {
-  pub target: super::CfgLabel,
-}
-
-// Action code 0x8f
-#[derive(Serialize, Deserialize, Debug, PartialEq, Eq)]
-#[serde(rename_all = "snake_case")]
-pub struct CfgTry {
-  pub r#try: crate::Cfg,
-  #[serde(skip_serializing_if = "Option::is_none")]
-  pub catch: Option<crate::Cfg>,
-  pub catch_target: crate::actions::r#try::CatchTarget,
-  #[serde(skip_serializing_if = "Option::is_none")]
-  pub finally: Option<crate::Cfg>,
-}
-
-#[derive(Serialize, Deserialize, Debug, PartialEq, Eq)]
-#[serde(rename_all = "snake_case")]
-pub struct CfgWith {
-  pub with: Option<crate::Cfg>,
+  pub target: Option<super::CfgLabel>,
 }